pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--strict-markers --doctest-modules --doctest-glob='*.rst' --ignore=setup.py"
doctest_optionflags = [
    "NORMALIZE_WHITESPACE",
    "IGNORE_EXCEPTION_DETAIL",
    "ELLIPSIS",
]
norecursedirs = [
    ".*",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    "htmlcov",
    "doc",
]
filterwarnings = []

[tool.ruff]
line-length = 205
target-version = "py39"
extend-exclude = ["docs"]

[tool.ruff.lint]
select = [
    "F",  # pyflakes
    "E",  # pycodestyle
    "W",  # pycodestyle
    "UP",  # pyupgrade
    "YTT",  # flake8-2020
    "B",  # flake8-bugbear
    "T10",  # flake8-debugger
    "C4",  # flake8-comprehensions
    "G",  # flake8-logging-format
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
]
ignore = []

[tool.black]
line-length = 205
target-version = ["py39"]
extend-exclude = "docs"

[tool.mypy]
warn_return_any = "True"
warn_unused_configs = "True"
strict_optional = "True"
ignore_missing_imports = "True"
disallow_any_unimported = "True"
check_untyped_defs = "True"
disallow_untyped_defs = "True"
no_implicit_optional = "True"
show_error_codes = "True"
warn_unused_ignores = "True"

# --- Buildsystem based on setuptools +  pyproject.toml : ---

[tool.setuptools_scm]
write_to = "_version.py"
write_to_template = "__version__ = \"{version}\"\n"

[tool.hatch.build.targets.sdist]
include = ["src/drift"]

[tool.hatch.build.targets.wheel]
include = ["src/drift"]

[tool.hatch.build.targets.wheel.sources]
"src/drift" = "drift"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools.packages.find]
where = ["src"]
include = ["drift"]
exclude = ["docs"]

[project]
name = "drift"
version = "1.0.6"
description = "Data Refresh and Intelligent Fast Training for Azure ML"
authors = [
    { name = "Simone DE SANTIS", email = "simone.desantis@amadeus.com" },
    { name = "Guillaume LECLERC", email = "guillaume.leclerc@amadeus.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-ml>=1.29.0,<2",
    "databricks-sdk>=0.38.0,<0.39",
    "deltalake>=0.21.0,<0.22",
    "mlflow-skinny>=3.3.2,<4",
    "mltable>=1.6.3,<2",
    "pydataio>=1.0.1",
    "pyspark==3.5.2",
]

[project.scripts]
spark_job = "drift.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
]

